Added TEMPLATE_MINIFIER_FILENAME_EXTENSIONS settings - allows minification only for particular template names
