+++ template_minifying_loader/loaders/app_directories.py
@@ -1,16 +1,7 @@
-from django.conf import settings
 from django.template.loaders.app_directories import Loader as AppDirectoriesLoader
+from ..mixins import TemplateMinifierMixin
 
-from ..utils import get_template_minifier_strip_function
 
-
-class Loader(AppDirectoriesLoader):
-
-    def get_contents(self, origin):
-        content = super().get_contents(origin)
-
-        if getattr(settings, 'TEMPLATE_MINIFIER', True):
-            content = get_template_minifier_strip_function()(content)
-
-        return content
+class Loader(TemplateMinifierMixin, AppDirectoriesLoader):
+    pass
