+++ template_minifying_loader/loaders/filesystem.py
@@ -1,16 +1,6 @@
-from django.conf import settings
 from django.template.loaders.filesystem import Loader as FilesystemLoader
-from django.template import Template
-
-from ..utils import get_template_minifier_strip_function
+from ..mixins import TemplateMinifierMixin
 
 
-class Loader(FilesystemLoader):
-
-    def get_contents(self, origin):
-        content = super().get_contents(origin)
-
-        if getattr(settings, 'TEMPLATE_MINIFIER', True):
-            content = get_template_minifier_strip_function()(content)
-
-        return content
+class Loader(TemplateMinifierMixin, FilesystemLoader):
+    pass
